backend/requirements.txt
flask
flask-cors
numpy
dimod
dwave-neal
waitress
